requirements.txt
Cython==0.20.1
Flask==0.10.1
Flask-Cors==1.9.0
Flask-Dropbox==0.3
Flask-PyMongo==0.3.0
Flask-Restless==0.13.1
Flask-SQLAlchemy==0.16
Flask-Script==2.0.5
Flask-Uploads==0.1.3
Flask-WTF==0.10.2
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.5.3
ProxyTypes==0.9
Pygments==1.6
SQLAlchemy==0.8.2
WTForms==2.0.1
Werkzeug==0.9.6
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
boto==2.32.1
certifi==14.05.14
decorator==3.4.0
dropbox==2.2.0
flask-mongoengine==0.7.1
gensim==0.9.1
google-api-python-client==1.2
gunicorn==17.5
html5lib==0.999
httplib2==0.9
itsdangerous==0.22
jabbapylib==0.3.8
matplotlib==1.4.0
mechanize==0.2.5
mercurial==3.1.1
mimerender==0.5.4
mock==1.0.1
mongoengine==0.8.7
mp3play==0.1.15
networkx==1.9
nose==1.3.3
numpy==1.9.0
ocropy==0.6
patsy==0.3.0
praw==2.1.18
psutil==2.1.1
py==1.4.24
pycurl==7.19.5
pygame==1.9.2a0
pyglet==1.1.4
pymongo==2.7.2
pyparsing==2.0.2
pytest==2.6.2
python-dateutil==2.1
python-loaders==0.2.3
python-mimeparse==0.1.4
python-tesseract==0.9
requests==1.2.3
scikit-image==0.10.1
scikit-learn==0.14.1
scipy==0.14.0
selenium==2.42.1
six==1.3.0
stevedore==0.15
tornado==3.2.1
twilio==3.6.7
untangle==1.1.0
update-checker==0.10
urllib3==1.9.1
uuid==1.30
vboxapi==1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.2
wsgiref==0.1.2
